# docs/openapi-backup.json
{
    "openapi":"3.0.0",
    "info":{
        "title": "UserHabitat",
        "description": "API REST for users and housing management",
        "version": "0.0.1"
    },
    "servers":[
        {
          "url": "https://api.userhabitat.com/v1",
          "description": "Main UserHabitat API server"
        }
    ],
    "paths":{
        "/users":{
            "get":{
                "summary":"Returns all users",
                "security": [
                    {
                        "OAuth2": ["read:users"]
                    }
                ],
                "responses":{
                    "200":{
                        "description":"List of users retrieved successfully",
                        "content":{
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/user"
                                    }
                                }
                            }
                        }
                    },
                    "401":{
                        "description": "Unauthorized",
                        "content": {
                          "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/error"
                                }
                            }
                        }
                    }
                }
            },
            "post":{
                "summary":"Creates a User",
                "security": [
                    {
                        "OAuth2": ["write:users"]
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/newUser"
                            }
                        }
                    }
                },
                "responses":{
                    "201":{
                        "description":"User created successfully",
                        "content":{
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/user"
                                }
                            }
                        }
                    },
                    "400":{
                        "description": "Invalid request. Missing or incorrect parameters",
                        "content": {
                          "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/error"
                                }
                            }
                        }
                    },
                    "401":{
                        "description": "Unauthorized",
                        "content": {
                          "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/error"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/users/{id}":{
            "get":{
                "summary":"Returns user by Id",
                "security": [
                    {
                        "OAuth2": ["read:users"]
                    }
                ],
                "responses":{
                    "200":{
                        "description":"User retrieved successfully",
                        "content":{
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/user"
                                }
                            }
                        }
                    },
                    "401":{
                        "description": "Unauthorized",
                        "content": {
                          "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/error"
                                }
                            }
                        }
                    },
                    "404":{
                        "description": "User not found",
                        "content": {
                          "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/error"
                                }
                            }
                        }
                    }
                }
            },
            "put":{
                "summary":"Updates a user completely",
                "security": [
                    {
                        "OAuth2": ["write:users"]
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/newUser"
                            }
                        }
                    }
                },
                "responses":{
                    "200":{
                        "description":"User updated successfully",
                        "content":{
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/user"
                                }
                            }
                        }
                    },
                    "400":{
                        "description": "Invalid request. Missing or incorrect parameters.",
                        "content": {
                          "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/error"
                                }
                            }
                        }
                    },
                    "401":{
                        "description": "Unauthorized",
                        "content": {
                          "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/error"
                                }
                            }
                        }
                    },
                    "404":{
                        "description": "User not found",
                        "content": {
                          "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/error"
                                }
                            }
                        }
                    }
                }
            },
            "patch":{
                "summary":"Updates a user partially",
                "security": [
                    {
                        "OAuth2": ["write:users"]
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Juan Gonzalez"
                                    }
                                },
                                "additionalProperties": false
                            }
                        }
                    }
                },
                "responses":{
                    "200":{
                        "description":"User partially updated successfully",
                        "content":{
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/user"
                                }
                            }
                        }
                    },
                    "400":{
                        "description": "Invalid request. Missing or incorrect parameters.",
                        "content": {
                          "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/error"
                                }
                            }
                        }
                    },
                    "401":{
                        "description": "Unauthorized",
                        "content": {
                          "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/error"
                                }
                            }
                        }
                    },
                    "404":{
                        "description": "User not found",
                        "content": {
                          "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/error"
                                }
                            }
                        }
                    }
                }
            },
            "delete":{
                "summary":"Deletes a user",
                "security": [
                    {
                        "OAuth2": ["write:users"]
                    }
                ],
                "responses":{
                    "204":{
                        "description":"User deleted successfully"
                    },
                    "400":{
                        "description": "The user cannot be deleted because it has associated houses",
                        "content": {
                          "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/error"
                                }
                            }
                        }
                    },
                    "401":{
                        "description": "Unauthorized",
                        "content": {
                          "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/error"
                                }
                            }
                        }
                    },
                    "404":{
                        "description": "User not found",
                        "content": {
                          "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/error"
                                }
                            }
                        }
                    }
                }
            },
            "parameters":[
                {
                    "name":"id",
                    "in":"path",
                    "description":"User Id",
                    "required":true,
                    "schema": {
                        "type": "string"
                    }
                }
            ]
        },
        "/users/{id}/houses":{
            "get":{
                "summary":"Returns all user's houses",
                "security": [
                    {
                        "OAuth2": ["read:houses"]
                    }
                ],
                "parameters":[
                    {
                        "name":"city",
                        "in":"query",
                        "description":"House city",
                        "required":false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name":"address",
                        "in":"query",
                        "description":"House address",
                        "required":false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name":"country",
                        "in":"query",
                        "description":"House country",
                        "required":false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses":{
                    "200":{
                        "description":"List of houses retrieved successfully",
                        "content":{
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/house"
                                    }
                                }
                            }
                        }
                    },
                    "401":{
                        "description": "Unauthorized",
                        "content": {
                          "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/error"
                                }
                            }
                        }
                    },
                    "404":{
                        "description": "User not found",
                        "content": {
                          "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/error"
                                }
                            }
                        }
                    }
                }
            },
            "post":{
                "summary":"Creates a user's house",
                "security": [
                    {
                        "OAuth2": ["write:houses"]
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/newHouse"
                            }
                        }
                    }
                },
                "responses":{
                    "201":{
                        "description":"House created successfully",
                        "content":{
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/house"
                                }
                            }
                        }
                    },
                    "400":{
                        "description": "Invalid request. Missing or incorrect parameters",
                        "content": {
                          "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/error"
                                }
                            }
                        }
                    },
                    "401":{
                        "description": "Unauthorized",
                        "content": {
                          "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/error"
                                }
                            }
                        }
                    },
                    "404":{
                        "description": "User not found",
                        "content": {
                          "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/error"
                                }
                            }
                        }
                    }
                }
            },
            "parameters":[
                {
                    "name":"id",
                    "in":"path",
                    "description":"User Id",
                    "required":true,
                    "schema": {
                        "type": "string"
                    }
                }
            ]
        },
...
    },
    "components":{
        "securitySchemes": {
            "OAuth2": {
                "type": "oauth2",
                "flows": {
                    "authorizationCode": {
                        "authorizationUrl": "https://auth.userhabitat.com/authorize",
                        "tokenUrl": "https://auth.userhabitat.com/token",
                        "scopes": {
                            "read:users": "Read access to users",
                            "write:users": "Write access to users",
                            "read:houses": "Read access to houses",
                            "write:houses": "Write access to houses"
                        }
                    }
                }
            }
        },
        "schemas":{
            "user":{
                "allOf":[
                    {
                        "$ref":"#/components/schemas/newUser"
                    },
                    {
                        "type": "object",
                        "properties":{
                            "id":{
                                "type":"string",
                                "example":"0000"
                            }
                        },
                        "required":["id"]
                    }
                ]
            },
            "newUser":{
                "type": "object",
                "properties":{
                    "name":{
                        "type":"string",
                        "example":"Juan Perez"
                    }
                },
                "required":["name"]
            },
            "house":{
                "allOf":[
                    {
                        "$ref":"#/components/schemas/newHouse"
                    },
                    {
                        "type": "object",
                        "properties":{
                            "id":{
                                "type":"string",
                                "example":"000"
                            }
                        },
                        "required":["id"]
                    }
                ]
            },
            "newHouse":{
                "type": "object",
                "properties": {
                    "address": {
                        "type": "string",
                        "example": "Calle Real 123"
                    },
                    "city": {
                        "type": "string",
                        "example": "Toledo"
                    },
                    "country": {
                        "type": "string",
                        "example": "Spain"
                    }
                },
                "required": ["address", "city", "country"]
            },
            "error": {
                "type": "object",
                "required": [
                    "code",
                    "message"
                ],
                "properties": {
                    "code": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "message": {
                        "type": "string"
                    }
                }
            }
        }
    }
} 